// src/infrastructure/grpc/generated/course/types/stats.ts
export interface MonthlyCoursesEnrollment {
  month: number;
  count: number;
}
// ...
export interface EnrollmentTrend {
  month: number;
  enrollments: number;
}

export interface GetInstructorCourseEnrollmentSummeryRequest {
  instructorId: string;
  courseId: string;
}

export interface InstructorCourseEnrollmentTrend {
  trend: EnrollmentTrend[];
}

export interface EnrollmentTrendStats {
  trend: EnrollmentTrend[];
}

export interface MonthlyCoursesEnrollmentStats {
  trend: EnrollmentTrend[];
}

export interface GetInstructorCoursesStatsRequest {
  instructorId: string;
}

export interface GetInstructorCoursesEnrollmentSummeryRequest {
  instructorId: string;
}

export interface GetMonthlyCoursesEnrollmentStatsRequest {
  year: string;
}
// ...
export const MonthlyCoursesEnrollment: MessageFns<MonthlyCoursesEnrollment> = {
  encode(
    message: MonthlyCoursesEnrollment,
    writer: BinaryWriter = new BinaryWriter(),
  ): BinaryWriter {
    if (message.month !== 0) {
      writer.uint32(8).int32(message.month);
    }
    if (message.count !== 0) {
      writer.uint32(16).int32(message.count);
    }
    return writer;
  },

  decode(
    input: BinaryReader | Uint8Array,
    length?: number,
  ): MonthlyCoursesEnrollment {
    const reader =
      input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseMonthlyCoursesEnrollment();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.month = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.count = reader.int32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): MonthlyCoursesEnrollment {
    return {
      month: isSet(object.month) ? globalThis.Number(object.month) : 0,
      count: isSet(object.count) ? globalThis.Number(object.count) : 0,
    };
  },

  toJSON(message: MonthlyCoursesEnrollment): unknown {
    const obj: any = {};
    if (message.month !== 0) {
      obj.month = Math.round(message.month);
    }
    if (message.count !== 0) {
      obj.count = Math.round(message.count);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<MonthlyCoursesEnrollment>, I>>(
    base?: I,
  ): MonthlyCoursesEnrollment {
    return MonthlyCoursesEnrollment.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<MonthlyCoursesEnrollment>, I>>(
    object: I,
  ): MonthlyCoursesEnrollment {
    const message = createBaseMonthlyCoursesEnrollment();
    message.month = object.month ?? 0;
    message.count = object.count ?? 0;
    return message;
  },
};
// ...
export const EnrollmentTrend: MessageFns<EnrollmentTrend> = {
  encode(
    message: EnrollmentTrend,
    writer: BinaryWriter = new BinaryWriter(),
  ): BinaryWriter {
    if (message.month !== 0) {
      writer.uint32(8).int32(message.month);
    }
    if (message.enrollments !== 0) {
      writer.uint32(16).int32(message.enrollments);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): EnrollmentTrend {
    const reader =
      input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseEnrollmentTrend();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.month = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.enrollments = reader.int32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): EnrollmentTrend {
    return {
      month: isSet(object.month) ? globalThis.Number(object.month) : 0,
      enrollments: isSet(object.enrollments)
        ? globalThis.Number(object.enrollments)
        : 0,
    };
  },

  toJSON(message: EnrollmentTrend): unknown {
    const obj: any = {};
    if (message.month !== 0) {
      obj.month = Math.round(message.month);
    }
    if (message.enrollments !== 0) {
      obj.enrollments = Math.round(message.enrollments);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<EnrollmentTrend>, I>>(
    base?: I,
  ): EnrollmentTrend {
    return EnrollmentTrend.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<EnrollmentTrend>, I>>(
    object: I,
  ): EnrollmentTrend {
    const message = createBaseEnrollmentTrend();
    message.month = object.month ?? 0;
    message.enrollments = object.enrollments ?? 0;
    return message;
  },
};
// ...
export const MonthlyCoursesEnrollmentStats: MessageFns<MonthlyCoursesEnrollmentStats> =
  {
    encode(
      message: MonthlyCoursesEnrollmentStats,
      writer: BinaryWriter = new BinaryWriter(),
    ): BinaryWriter {
      for (const v of message.trend) {
        EnrollmentTrend.encode(v!, writer.uint32(10).fork()).join();
      }
      return writer;
    },

    decode(
      input: BinaryReader | Uint8Array,
      length?: number,
    ): MonthlyCoursesEnrollmentStats {
      const reader =
        input instanceof BinaryReader ? input : new BinaryReader(input);
      const end = length === undefined ? reader.len : reader.pos + length;
      const message = createBaseMonthlyCoursesEnrollmentStats();
      while (reader.pos < end) {
        const tag = reader.uint32();
        switch (tag >>> 3) {
          case 1: {
            if (tag !== 10) {
              break;
            }

            message.trend.push(EnrollmentTrend.decode(reader, reader.uint32()));
            continue;
          }
        }
        if ((tag & 7) === 4 || tag === 0) {
          break;
        }
        reader.skip(tag & 7);
      }
      return message;
    },

    fromJSON(object: any): MonthlyCoursesEnrollmentStats {
      return {
        trend: globalThis.Array.isArray(object?.trend)
          ? object.trend.map((e: any) => EnrollmentTrend.fromJSON(e))
          : [],
      };
    },

    toJSON(message: MonthlyCoursesEnrollmentStats): unknown {
      const obj: any = {};
      if (message.trend?.length) {
        obj.trend = message.trend.map((e) => EnrollmentTrend.toJSON(e));
      }
      return obj;
    },

    create<I extends Exact<DeepPartial<MonthlyCoursesEnrollmentStats>, I>>(
      base?: I,
    ): MonthlyCoursesEnrollmentStats {
      return MonthlyCoursesEnrollmentStats.fromPartial(base ?? ({} as any));
    },
    fromPartial<I extends Exact<DeepPartial<MonthlyCoursesEnrollmentStats>, I>>(
      object: I,
    ): MonthlyCoursesEnrollmentStats {
      const message = createBaseMonthlyCoursesEnrollmentStats();
      message.trend =
        object.trend?.map((e) => EnrollmentTrend.fromPartial(e)) || [];
      return message;
    },
  };
// ...
export const GetMonthlyCoursesEnrollmentStatsRequest: MessageFns<GetMonthlyCoursesEnrollmentStatsRequest> =
  {
    encode(
      message: GetMonthlyCoursesEnrollmentStatsRequest,
      writer: BinaryWriter = new BinaryWriter(),
    ): BinaryWriter {
      if (message.year !== "") {
        writer.uint32(10).string(message.year);
      }
      return writer;
    },

    decode(
      input: BinaryReader | Uint8Array,
      length?: number,
    ): GetMonthlyCoursesEnrollmentStatsRequest {
      const reader =
        input instanceof BinaryReader ? input : new BinaryReader(input);
      const end = length === undefined ? reader.len : reader.pos + length;
      const message = createBaseGetMonthlyCoursesEnrollmentStatsRequest();
      while (reader.pos < end) {
        const tag = reader.uint32();
        switch (tag >>> 3) {
          case 1: {
            if (tag !== 10) {
              break;
            }

            message.year = reader.string();
            continue;
          }
        }
        if ((tag & 7) === 4 || tag === 0) {
          break;
        }
        reader.skip(tag & 7);
      }
      return message;
    },

    fromJSON(object: any): GetMonthlyCoursesEnrollmentStatsRequest {
      return { year: isSet(object.year) ? globalThis.String(object.year) : "" };
    },

    toJSON(message: GetMonthlyCoursesEnrollmentStatsRequest): unknown {
      const obj: any = {};
      if (message.year !== "") {
        obj.year = message.year;
      }
      return obj;
    },

    create<
      I extends Exact<DeepPartial<GetMonthlyCoursesEnrollmentStatsRequest>, I>,
    >(base?: I): GetMonthlyCoursesEnrollmentStatsRequest {
      return GetMonthlyCoursesEnrollmentStatsRequest.fromPartial(
        base ?? ({} as any),
      );
    },
    fromPartial<
      I extends Exact<DeepPartial<GetMonthlyCoursesEnrollmentStatsRequest>, I>,
    >(object: I): GetMonthlyCoursesEnrollmentStatsRequest {
      const message = createBaseGetMonthlyCoursesEnrollmentStatsRequest();
      message.year = object.year ?? "";
      return message;
    },
  };
// ...
type Builtin =
  | Date
  | Function
  | Uint8Array
  | string
  | number
  | boolean
  | undefined;

export type DeepPartial<T> = T extends Builtin
  ? T
  : T extends globalThis.Array<infer U>
    ? globalThis.Array<DeepPartial<U>>
    : T extends ReadonlyArray<infer U>
      ? ReadonlyArray<DeepPartial<U>>
      : T extends {}
        ? { [K in keyof T]?: DeepPartial<T[K]> }
        : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin
  ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & {
      [K in Exclude<keyof I, KeysOfUnion<P>>]: never;
    };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}